@@ -4,9 +4,11 @@
   "description": "Angular 2 bindings for Redux",
   "main": "./lib/index.js",
   "scripts": {
-    "build": "rm -rf lib && `npm bin`/babel src --out-dir lib",
-    "test": "mocha --compilers js:babel/register --recursive"
+    "postinstall": "tsd install",
+    "build": "rm -rf lib; tsc",
+    "test": "mocha --compilers js:babel/register --recursive --require ts-node/register"
   },
+  "typings": "lib/",
   "repository": {
     "type": "git",
     "url": "git+ssh://git@github.com/wbuchwalter/ng2-redux.git"
@@ -23,15 +25,19 @@
   },
   "homepage": "https://github.com/wbuchwalter/ng2-redux#readme",
   "devDependencies": {
+    "angular2": "^2.0.0-alpha.53",
     "babel": "^5.8.12",
     "babel-loader": "^5.3.2",
     "expect": "^1.8.0",
     "mocha": "^2.2.5",
-    "sinon": "^1.16.1",
-    "webpack": "^1.10.5",
-    "angular2": "^2.0.0-alpha.53",
     "react": "^0.13.3",
-    "redux-devtools": "^2.1.0"
+    "redux-devtools": "^2.1.0",
+    "rxjs": "^5.0.0-beta.1",
+    "sinon": "^1.16.1",
+    "ts-node": "^0.5.5",
+    "tsd": "^0.6.5",
+    "typescript": "^1.7.5",
+    "webpack": "^1.12.11"
   },
   "peerDependencies": {
     "redux": "^2.0.0 || ^3.0.0",
@@ -39,6 +45,7 @@
   },
   "dependencies": {
     "invariant": "^2.1.0",
-    "lodash": "^3.10.1"
+    "lodash": "^3.10.1",
+    "reflect-metadata": "^0.1.3"
   }
 }
