@@ -1,52 +1,80 @@
-// Type definitions for Redux v1.0.0
+// Type definitions for Redux v3.0.5
 // Project: https://github.com/rackt/redux
-// Definitions by: William Buchwalter <https://github.com/wbuchwalter/>, Vincent Prouillet <https://github.com/Keats/>
+// Definitions by: William Buchwalter <https://github.com/wbuchwalter/>, Vincent Prouillet <https://github.com/Keats/>, Michael Bennett <https://github.com/bennett000/>
 // Definitions: https://github.com/borisyankov/DefinitelyTyped
+
+declare let Redux: Redux.ReduxStatic;
+
+declare module 'redux' {
+  export = Redux;
+}
 
 declare module Redux {
 
-    interface ActionCreator extends Function {
-        (...args: any[]): any;
-    }
+  interface Action {
+    type: string;
+    payload?: any;
+    error?: boolean;
+    meta?: any;
+  }
 
-    interface Reducer extends Function {
-        (state: any, action: any): any;
-    }
+  interface Map<T> {
+    [id: string]: T;
+  }
 
-    interface Dispatch extends Function {
-        (action: any): any;
-    }
+  interface ActionCreator {
+    (...args: any[]): Action;
+  }
 
-    interface StoreMethods {
-        dispatch: Dispatch;
-        getState(): any;
-    }
+  interface Listener {
+    (): void;
+  }
 
+  interface Unsubscribe {
+    (): void;
+  }
 
-    interface MiddlewareArg {
-        dispatch: Dispatch;
-        getState: Function;
-    }
+  interface Reducer<T> {
+    (state: T, action: Action): T;
+  }
 
-    interface Middleware extends Function {
-        (obj: MiddlewareArg): Function;
-    }
+  interface Dispatch {
+    (action: Action): Action;
+  }
 
-    class Store {
-        getReducer(): Reducer;
-        replaceReducer(nextReducer: Reducer): void;
-        dispatch(action: any): any;
-        getState(): any;
-        subscribe(listener: Function): Function;
-    }
+  interface PartialDispatch {
+    (): Action;
+  }
 
-    function createStore(reducer: Reducer, initialState?: any): Store;
-    function bindActionCreators<T>(actionCreators: T, dispatch: Dispatch): T;
-    function combineReducers(reducers: any): Reducer;
-    function applyMiddleware(...middleware: Middleware[]): Function;
-    function compose<T extends Function>(...functions: Function[]): T;
-}
+  interface MiddlewareArg<T> {
+    dispatch: Dispatch;
+    getState: T;
+  }
 
-declare module "redux" {
-    export = Redux;
+  interface Middleware<T> {
+    (obj: MiddlewareArg<T>): Function;
+  }
+
+  interface Store<T> {
+    replaceReducer(nextReducer: Reducer<T>): void;
+    dispatch(action: Action): Action;
+    getState(): T;
+    subscribe(listener: Listener): Unsubscribe;
+  }
+
+  interface CreateStore<T> {
+    (reducer: Reducer<T>, initialState?: T): Store<T>;
+  }
+
+  interface ReduxStatic {
+    createStore<T>(reducer: Reducer<T>, initialState?: T): Store<T>;
+    bindActionCreators<T extends Map<ActionCreator>,
+      TP extends Map<PartialDispatch>>(actionCreators: T,
+                                       dispatch: Dispatch): TP;
+    bindActionCreators(actionCreator: ActionCreator,
+                       dispatch: Dispatch): ActionCreator;
+    combineReducers<T>(reducers: Map<Reducer<any>>): Reducer<T>;
+    applyMiddleware<T>(...middlewares: Middleware<T>[]): CreateStore<T>;
+    compose<T extends Function>(...functions: Function[]): T;
+  }
 }
